cookbook/pocketflow-palantir-osdk/requirements.txt
pocketflow
openai>=1.0.0
streamlit>=1.28.0
plotly>=5.18.0
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
